move `synchronize.ts` back out into `test-context/`
