+++ projects/ng-dev/src/lib/test-context/angular-context/angular-context.ts
@@ -20,7 +20,7 @@ import {
 import { assert, convertTime } from '@s-libs/js-core';
 import { clone, forOwn, isFunction } from '@s-libs/micro-dash';
 import { FakeAsyncHarnessEnvironment } from './fake-async-harness-environment';
-import { Synchronized } from './synchronize';
+import { Synchronized } from '../synchronize';
 
 /** @hidden */
 export function extendMetadata(
