+++ projects/ng-dev/src/lib/test-context/angular-context/fake-async-harness-environment.spec.ts
@@ -2,7 +2,7 @@ import { ContentContainerComponentHarness } from '@angular/cdk/testing';
 import { Component } from '@angular/core';
 import { MatButtonHarness } from '@angular/material/button/testing';
 import { ComponentContextNext } from '../component-context';
-import { Synchronized } from './synchronize';
+import { Synchronized } from '../synchronize';
 
 @Component({
   template: `
