+++ projects/ng-dev/src/lib/test-context/angular-context/index.ts
@@ -1,2 +1,1 @@
 export { AngularContext } from './angular-context';
-export { Synchronized } from './synchronize';
