+++ projects/ng-dev/src/lib/test-context/index.ts
@@ -1,2 +1,3 @@
 export * from './angular-context';
 export * from './component-context';
+export { Synchronized } from './synchronize';
